refac: Regions를 삭제하고, Regions에 있던 로직을 LegalArea에 옮김. 또한, 추상 클래스였던 Area를 인터페이스로 수정
